requirements.txt
fastmcp>=0.1.0
pymongo>=4.6.0
pydantic>=2.0.0
python-dotenv>=1.0.0
dnspython>=2.4.0 